Refactor StackedCalculator to separate file
Move StackedCalculator to src/et_models/stackedcalc.jl for better
separation of concerns - it's a generic utility for combining
calculators, independent of ETACE-specific code.
